Add files via upload
Added scoring system and collision detection that ends the game. Game is now functional, possible room for improvement includes adding the ability to restart the game, to skip or justify the "pause" before you can shoot the cannon, a menu function, a leader board function. In addition, code is missing documentation.